+++ src/frontend/babel.config.js
@@ -1,0 +1,20 @@
+module.exports = {
+  plugins: [
+    [
+      'react-intl',
+      {
+        messagesDir: './i18n',
+      },
+    ],
+  ],
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        forceAllTransforms: true,
+        useBuiltIns: 'usage',
+      },
+    ],
+    'react',
+  ],
+};